Implement the most horrendous return system ever
It works, at least, and does so by putting a return signal in the Err()
and at that signal reading out the `return_value` field in the
Enviroment struct. Welp, good enough for now. We are learning here
remember. (Cries in spaghetti.)

functions.lox shows what we are capable of now.
fib.lox is the recursive Fibonacci ordeal. Runs in 387ms in release
mode, currently.